Config refactor; Still we need to apply: https://github.com/lorenwest/node-config/wiki/Webpack-Usage
