+++ config/local.json
@@ -1,0 +1,32 @@
+{
+  "elasticsearch": {
+    "host": "localhost:8080/api/catalog",
+    "index": "vue_storefront_catalog"
+  },
+  "orders": {
+    "endpoint": "http://localhost:8080/api/order"
+  },
+  "users": {
+    "endpoint": "http://localhost:8080/api/user"
+  },
+  "stock": {
+    "endpoint": "http://localhost:8080/api/stock"
+  },  
+  "images": {
+    "baseUrl": "https://demo.vuestorefront.io/img/"
+  },
+  "demomode": true,
+  "tax": {
+    "defaultCountry": "PL",
+    "defaultRegion": "",
+    "calculateServerSide": true
+  },  
+  "i18n": {
+    "defaultCountry": "US",
+    "defaultLanguage": "EN"
+  },
+  "mailchimp": {
+    "endpoint": "http://localhost:8080/api/ext/mailchimp-subscribe/subscribe"
+  },
+  "registeredExtensions": ["custom_extension", "mailchimp-subscribe"]
+}
