+++ installer.js
@@ -12,8 +12,8 @@ const isEmptyDir = require('empty-dir')
 const commandExists = require('command-exists')
 
 const SAMPLE_DATA_PATH = 'var/magento2-sample-data'
-const TARGET_CONFIG_FILE = 'src/config.json'
-const SOURCE_CONFIG_FILE = 'src/config.example.json'
+const TARGET_CONFIG_FILE = 'config/default.json'
+const SOURCE_CONFIG_FILE = 'config/local.json'
 
 const STOREFRONT_GIT_URL = 'https://github.com/DivanteLtd/vue-storefront'
 const STOREFRONT_BACKEND_GIT_URL = 'https://github.com/DivanteLtd/vue-storefront-api'
@@ -196,7 +196,7 @@ class Backend extends Abstract {
   }
 
   /**
-   * Creating backend src/config.json
+   * Creating backend config/local.json
    *
    * @returns {Promise}
    */
@@ -321,7 +321,7 @@ class Storefront extends Abstract {
   }
 
   /**
-   * Creating storefront src/config.json
+   * Creating storefront config/local.json
    *
    * @returns {Promise}
    */
