+++ package.json
@@ -37,6 +37,7 @@
     "axios": "^0.13.1",
     "bodybuilder": "^2.2.1",
     "command-exists": "^1.2.2",
+    "config": "^1.29.1",
     "cross-env": "^3.1.4",
     "elastic-builder": "^1.1.0",
     "elasticsearch": "^13.3.1",
