+++ src/config/index.js
@@ -1,0 +1,17 @@
+const Config = require('config')
+
+const ConfigPlugin = {
+  install (Vue) {
+    if (!Vue.prototype.$config) {
+      Object.defineProperties(Vue.prototype, {
+        $config: {
+          get: function () {
+            return Config
+          }
+        }
+      })
+    }
+  }
+}
+
+export { Config as default, ConfigPlugin }
