+++ src/store/modules/user.js
@@ -1,5 +1,5 @@
 import * as types from '../mutation-types'
-import config from '../../config.json'
+import { Config as config } from 'src/config'
 import EventBus from 'src/event-bus'
 
 // initial state
